LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/module/evaluation-order-3-dynamic-expected.txt


FAIL Test evaluation order of modules assert_array_equals: expected property 3 to be "microtask-3.1" but got "script-load" (expected array ["step-3.1-1", "step-3.1-2", "step-3.1-3", "microtask-3.1", "script-load", "step-3.2-1", "step-3.2-2", "microtask-3.2", "import-catch", "error"] got ["step-3.1-1", "step-3.1-2", "step-3.1-3", "script-load", "microtask-3.1", "step-3.2-1", "step-3.2-2", "microtask-3.2", "import-catch", "error"])
